panel admin kamel ta ghabl as emkan kharid

+++ app/Http/Controllers/CourseController.php
@@ -7,33 +7,38 @@ use App\Models\Category;
 use App\Models\Course;
 use App\Models\Video;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
 
 class CourseController extends Controller
 {
     public function index()
     {
-        $courses = Course::orderBy('created_at','desc')->get();
+        $courses = Course::orderBy('created_at', 'desc')->get();
         return view('course.index')
-            ->with('courses',$courses);
+            ->with('courses', $courses);
     }
+
     public function display(Course $course)
     {
         $videos = $course->videos;
         return view('course.display')
-            ->with('course',$course)
-            ->with('videos',$videos);
+            ->with('course', $course)
+            ->with('videos', $videos);
     }
+
     public function create()
     {
-        $categories =Category::with('parent')->whereNotNull('parent_id')->get();
+        $categories = Category::with('parent')->whereNotNull('parent_id')->get();
         $main = Category::with('parent')->whereNull('parent_id')->get();
         return view('course.create')
-            ->with('main',$main)
-            ->with('categories',$categories);
+            ->with('main', $main)
+            ->with('categories', $categories);
     }
-    public function save(StoreCourse $request,Course $course)
+
+    public function save(StoreCourse $request, Course $course)
     {
         $course->user_id = $request->user_id;
         $course->title = $request->title;
@@ -41,14 +46,14 @@ class CourseController extends Controller
         $course->price = $request->price;
 
         if ($request->hasFile('avatar')) {
-            $avatarPath = Storage::disk('uploads')->put('/image/' . $request->title, $request->file('avatar'));
+            $avatarPath = Storage::disk('uploads')->put('/courses/' . $request->title, $request->file('avatar'));
             $course->avatar = 'http://localhost:8000/' . $avatarPath;
         } else {
             $course->avatar = null;
         }
 
         if ($request->hasFile('source')) {
-            $sourcePath = Storage::disk('uploads')->put('/image/' . $request->title, $request->file('source'));
+            $sourcePath = Storage::disk('uploads')->put('/courses/' . $request->title, $request->file('source'));
             $course->source = 'http://localhost:8000/' . $sourcePath;
         } else {
             $course->source = null;
@@ -57,21 +62,35 @@ class CourseController extends Controller
         $course->save();
         $course->categories()->attach($request->input('category'));
         return redirect('/admin/course')
-            ->with('success','دوره با موفقیت ذخیره شد');
+            ->with('success', 'دوره با موفقیت ذخیره شد');
 
     }
+
     public function edit(Course $course)
     {
-        $categories =Category::with('parent')->whereNotNull('parent_id')->get();
+        $categories = Category::with('parent')->whereNotNull('parent_id')->get();
         $main = Category::with('parent')->whereNull('parent_id')->get();
         $cats = $course->categories;
-        return view('course.edit')
-            ->with('main',$main)
-            ->with('categories',$categories)
-            ->with('course',$course)
-            ->with('cats',$cats);
+        if (Auth::user()->id == $course->user_id or Auth::user()->level === 'مدیر') {
+            return view('course.edit')
+                ->with('main', $main)
+                ->with('categories', $categories)
+                ->with('course', $course)
+                ->with('cats', $cats);
+        } else {
+            return redirect('/admin/course')->with('warning', 'دسترسی مجاز نیست');
+        }
+
     }
-    public function update(StoreCourse $request,Course $course)
+
+    public function getFile($file)
+    {
+        $file = ltrim($file, "p://localhost:8000/");
+        $file = str_replace('/', '\\', $file);
+        return $file;
+    }
+
+    public function update(StoreCourse $request, Course $course)
     {
         $course->user_id = $request->user_id;
         $course->title = $request->title;
@@ -80,57 +99,41 @@ class CourseController extends Controller
         $course->categories()->sync($request->input('category'));
 
         if ($request->hasFile('avatar')) {
-            $avatarPath = Storage::disk('uploads')->put('/image/' . $request->title, $request->file('avatar'));
+            unlink(dirname(storage_path()) . '\\public\\courses\\' . str_replace('/', '\\', trim($request->oldAvatar, 'http://localhost:8000/courses/')));
+            $avatarPath = Storage::disk('uploads')->put('/courses/' . $request->title, $request->file('avatar'));
             $course->avatar = 'http://localhost:8000/' . $avatarPath;
         } else {
-            $course->avatar = null;
+            $course->avatar = $request->oldAvatar;
         }
-
         if ($request->hasFile('source')) {
-            $sourcePath = Storage::disk('uploads')->put('/image/' . $request->title, $request->file('source'));
+            unlink(dirname(storage_path()) . '\\public\\courses\\' . str_replace('/', '\\', trim($request->oldSource, 'http://localhost:8000/courses/')));
+            $sourcePath = Storage::disk('uploads')->put('/courses/' . $request->title, $request->file('source'));
             $course->source = 'http://localhost:8000/' . $sourcePath;
         } else {
-            $course->source = null;
+            $course->source = $request->oldSource;
         }
         $course->save();
-        return redirect('/admin/course')->with('success','دوره با موفقیت ویرایش شد');
+        return redirect('/admin/course')->with('success', 'دوره با موفقیت ویرایش شد');
     }
-
-
 
 
     public function destroy(Course $course)
     {
-        $courseVideos = $course->videos;
-        if (count($courseVideos) > 0)
-        {
-            foreach ($courseVideos as $video)
-            {
-                $video->delete();
+        if (Auth::user()->id == $course->user_id or Auth::user()->level === 'مدیر') {
+            $courseVideos = $course->videos;
+            if (count($courseVideos) > 0) {
+                foreach ($courseVideos as $video) {
+                    $video->delete();
+                }
             }
+            $cat = DB::table('category_course')->where('course_id', $course->id)->delete();
+            File::deleteDirectory(dirname(storage_path()) . '\\public\\courses\\' . $course->title);
+            $course->delete();
+            return redirect('/admin/course')->with('success', 'دوره با موفقیت حذف شد');
+        } else {
+            return redirect('/admin/course')->with('warning', 'دسترسی مجاز نیست');
         }
-        $cat = DB::table('category_course')->where('course_id',$course->id)->delete();
-        Storage::delete('public/image/'.$course->title);
-        $course->delete();
-        return redirect('/admin/course')->with('success','دوره با موفقیت حذف شد');
     }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 
     public function show($slug1, $slug2 = null)
@@ -138,24 +141,28 @@ class CourseController extends Controller
         if ($slug2 == null) {
             $slug1 = str_replace('-', ' ', $slug1);
             $course = Course::where('title', $slug1)->firstOrFail();
+            $categories = $course->categories;
+            $name = $course->user->name;
             $videos = $course->videos;
             return view('course.show')
                 ->with('videos', $videos)
+                ->with('name', $name)
+                ->with('categories', $categories)
                 ->with('course', $course);
         } else {
             $slug1 = str_replace('-', ' ', $slug1);
             $slug2 = str_replace('-', ' ', $slug2);
-            $video = Video::where('title', $slug2)->firstOrFail();
+            //
+            //return $video;
             $course = Course::where('title', $slug1)->firstOrFail();
+            $courseId = $course->id;
+            $video = Video::where('title', $slug2)->where('course_id', $courseId)->firstOrFail();
             $videos = $course->videos;
-            if ($video->course_id == $course->id) {
-                return view('video.show')
-                    ->with('videos', $videos)
-                    ->with('course', $course)
-                    ->with('video', $video);
-            } else {
-                return abort(404);
-            }
+            return view('video.show')
+                ->with('videos', $videos)
+                ->with('course', $course)
+                ->with('video', $video);
+
 
         }
     }
